指定Notify中的通知回调函数
