+++ conf/dev/demo.yml
@@ -1,0 +1,54 @@
+application:
+  name: demo-application
+  debug: true
+
+server:
+  addr: 0.0.0.0:13360
+
+
+include: # 其他配置子文件，复用的配置文件
+  data: data.yml
+
+alter:
+  users:
+    - lvfei
+    - fangwendong
+    - xupeng
+
+
+cache:
+  normal:
+    redis: demo
+    lru-enable: true
+    lru-max-size: 50000
+    lru-expire-duration: 1m
+
+
+log:
+  error:
+    logger:
+      filename: log/error.log
+      maxsize: 100
+      maxage: 7
+      maxbackups: 10
+      compress: true
+    interval: 1h
+    level: debug
+  access:
+    logger:
+      filename: log/access.log
+      maxsize: 100
+      maxage: 7
+      maxbackups: 10
+      compress: true
+    interval: 1h
+    level: debug
+  access_error:
+    logger:
+      filename: log/access_error.log
+      maxsize: 100
+      maxage: 7
+      maxbackups: 10
+      compress: true
+    interval: 1h
+    level: debug
