--- pkg/cmd/demo/route.go
@@ -1,4 +1,0 @@
-package main
-
-func registerRouter() {
-}
